+++ Marindeck/View/Core/ViewController.swift
@@ -8,7 +8,6 @@
 import UIKit
 import WebKit
 import SafariServices
-import LocalAuthentication
 
 import Keys
 import Optik
@@ -77,8 +76,8 @@ class ViewController: UIViewController, UIScrollViewDelegate, UIAdaptivePresenta
     }
     
     override var keyCommands: [UIKeyCommand]? {
-        return [
-            .init(title: L10n.ActionButton.Tweet.title, action: #selector(self.tweetPressed), input: "n", modifierFlags: [.command])
+        [
+            .init(title: L10n.ActionButton.Tweet.title, action: #selector(tweetPressed), input: "n", modifierFlags: [.command])
         ]
     }
 
@@ -156,50 +155,7 @@ class ViewController: UIViewController, UIScrollViewDelegate, UIAdaptivePresenta
 //             scrollView.pinchGestureRecognizer?.isEnabled = false
 //    }
 
-    // 生体認証
-    func checkBiometrics() {
-        let isUseBiometrics = UserDefaults.standard.bool(forKey: UserDefaultsKey.isUseBiometrics)
-        if isUseBiometrics {
-            if canUseBiometrics() {
-                isMainDeckViewLock = true
-                mainDeckView.isHidden = true
-                let context = LAContext()
-                let reason = "ロックを解除"
-                let backBlackView = UIView(frame: view.bounds)
-                backBlackView.backgroundColor = .black
-                view.addSubview(backBlackView)
-                context.evaluatePolicy(.deviceOwnerAuthentication, localizedReason: reason) { (success, evaluateError) in
-                    if success {
-                        DispatchQueue.main.async { [unowned self] in
-                            self.isMainDeckViewLock = false
-                            self.mainDeckView.isHidden = false
-                            self.tweetFloatingBtn.isHidden = false
-                            UIView.animate(withDuration: 0.3, animations: {
-                                backBlackView.alpha = 0
-                            }, completion: { _ in
-                                backBlackView.removeFromSuperview()
-                            })
-                        }
-                    } else {
-                        DispatchQueue.main.async {
-                            let errorLabel = UILabel(frame: backBlackView.bounds)
-                            errorLabel.textAlignment = .center
-                            errorLabel.text = "認証に失敗しました。"
-                            errorLabel.textColor = .white
-                            backBlackView.addSubview(errorLabel)
-                        }
-                        guard let error = evaluateError as NSError? else {
-                            print("Error")
-                            return
-                        }
-                        print("\(error.code): \(error.localizedDescription)")
-                    }
-                }
-            } else {
-                // 生体認証をオンにしているが、許可されていない。
-            }
-        }
-    }
+
 
     // ツイートボタンタップ痔の動作
     @objc func tweetPressed() {
@@ -591,7 +547,6 @@ extension ViewController: ImageViewerDelegate {
     func imageViewerDidDisplayImage(at index: Int) {
         imagePreviewSelectedIndex = index
         setPreviewImagePosition()
-        print("imageVIewerDidDisplafmeifimageeee")
     }
